refactor(items): 重构打包材料命名并调整配方布局
- 将 package_material_x1 重命名为 package_material
- 更新所有配方中材料引用从 package_material_x1 到 package_material
- 调整配方图案布局，将打包材料放置于中间行左右两端
- 更新配方文档以反映新的配方图案和材料名称
- 修复GUI能量显示中的整数溢出问题
- 添加压板卡帮助界面的箭头符号支持

+++ src/main/java/cn/sd/jrz/alltheimbaium/gui/InstantFurnaceScreen.java
@@ -170,7 +170,8 @@ public class InstantFurnaceScreen extends AbstractContainerScreen<InstantFurnace
         int max = Math.max(1, this.menu.getMaxEnergy());
         int energy = Math.max(0, Math.min(max, this.menu.getEnergy()));
         if (energy > 0) {
-            int fill = ENERGY_W * energy / max;
+            // 必须先用 long 相乘：MAX_ENERGY 是 20 亿，ENERGY_W × energy 会溢出 int
+            int fill = (int) ((long) ENERGY_W * energy / max);
             guiGraphics.fill(trackLeft, trackTop, trackLeft + fill, trackTop + ENERGY_H, ENERGY_FILL_COLOR);
         }
     }
